backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.34.0
elevenlabs>=1.0.0

